+++ src/java/logicsim/OffDelay.java
@@ -56,7 +56,7 @@ public class OffDelay extends Gate {
   public boolean showProperties(Component frame) {
         String h = (String)JOptionPane.showInputDialog(frame, I18N.getString("GATE_TURNOFFDELAY_TIME"), I18N.getString("GATE_TURNOFFDELAY"),
                                                     JOptionPane.QUESTION_MESSAGE, null,null, Integer.toString((int)delayTime));
-    if (h!=null && h.length()>0) delayTime=new Integer(h).intValue();
+    if (h!=null && h.length()>0) delayTime= new Integer(h);
     return true;
   }
